fix(content-manager): model MCP dynamic zones and tighten component id/aggregate schemas
Addresses the third review round on the MCP draft-write schemas.

Dynamic zones were modelled as `z.array(z.any())`, so neither `__component` nor
the target component's required fields were validated. Dynamic-zone entries go
through the same `updateOrCreateComponent` dispatch as ordinary components, so
an id-less entry CREATEs a replacement row and deletes the old one — meaning an
entry could replace a valid component while omitting a required leaf. Recursive
entity validation applies update semantics and only checks `notNull`, which an
absent key satisfies, so on a non-D&P model the incomplete component went live.
Entries are now a `__component`-discriminated union that delegates to
`buildComponentInputSchema`, inheriting the id-patch / id-less-create split.

Component `id` used a bare `z.coerce.number()`, which turns `null`, `false` and
`""` into `0` and `true` into `1`, and accepts fractional and negative values —
each silently selecting a row rather than failing. It now accepts only positive
integers and digit strings. This advertises a two-member `anyOf` instead of a
plain `number`, which is the accurate contract: coercion claimed `number` while
accepting `null`.

Required repeatable components and dynamic zones are no longer relaxed on draft
creates. `createComponentValidator` (repeatable branch) and `createDzValidator`
pass a hard-coded `required: true` to `addRequiredValidation` regardless of
`isDraft`, and `addDefault` only substitutes `[]` for non-required aggregates —
so omitting the key on create throws "<field> must be defined" even for a draft.
Relaxing it advertised a write the server rejects. Aggregate projection is now
separate from the scalar one (`applyAggregateRequired`) and relaxes only on
update, where `notNull()` genuinely accepts an absent key.

The component-replacement API tests now run against a non-D&P article. They
previously wrote the shared D&P article as a draft, where a missing required
leaf is intentionally valid, so they demonstrated draft leniency rather than the
replacement hazard. With draft & publish off the writes are published and the
integrity gap is what the assertions actually pin down. Adds a dynamic-zone case
proving an id-less entry persists a missing required field on live content.
